- исправление мелких багов - добавление опшинал енд поинтов - добавление урл в книгах - добавление работы урл в создание видах - тестовый джсон для создание книг - добавление нового соурса

+++ src/main/java/club/ttg/dnd5/model/book/Source.java
@@ -17,7 +17,7 @@ public class Source extends TimestampedEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private short page;
+    private int page;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "book_info_id", referencedColumnName = "sourceAcronym")
